refactor: split notify-hook.js into layered sub-modules (issue #177)
Extracts the 1577-line monolith into 9 focused modules under
scripts/notify-hook/, each with a single clear responsibility:

  utils.js           – pure helpers (asNumber, safeString, clampPct, …)
  payload-parser.js  – payload field extraction (tokens, quota, messages)
  state-io.js        – state file I/O and normalization
  process-runner.js  – child-process helper (runProcess)
  log.js             – structured tmux-hook event logging
  auto-nudge.js      – stall-pattern detection and auto-nudge
  linked-sync.js     – linked ralph/team terminal sync
  tmux-injection.js  – tmux prompt injection and pane resolution
  team-leader-nudge.js – leader mailbox/staleness nudge
  team-worker.js     – worker heartbeat and all-workers-idle notification

notify-hook.js becomes a thin orchestrator (~250 lines) that imports
from these modules. All existing integration tests continue to pass.

Also fixes the dynamic readdir() wrapper at the bottom of the original
file: readdir is now imported statically in state-io.js instead of
re-imported via a dynamic import on every call.

Adds 33 new unit tests (src/hooks/__tests__/notify-hook-modules.test.ts)
that directly import and test pure functions from the sub-modules —
something not possible with the previous monolithic structure.
